requirements.txt
boto3==1.35.21    
fastapi==0.112.0
uvicorn==0.30.6
logging==0.4.9.6
awscli==1.34.21
asyncio == 3.4.3
aiohttp == 3.10.5
paramiko == 3.5.0
numpy == 2.1.2
matplotlib == 3.9.2
mysql-connector-python == 9.1.0
requests == 2.32.3
httpx == 0.28.0